nua_cli/pyproject.toml
[tool.poetry]
name = "nua_cli"
version = "0.1.0-1655995893"
description = "Some nua_cli implementation with Typer"
authors = ["Jerome Dumonteil <jd@abilian.com>"]
license = "AGPL"
readme = "README.md"

[tool.poetry.scripts]
nua = "nua_cli.main:app"

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
wheel = "^0.37.1"
setuptools = "^62.3.2"
typer = {extras = ["all"], version = "^0.4.1"}
py = "^1.11.0"
pyzmq = "^23.1.0"
tinyrpc = "^1.1.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
# Linting
flake8 = "^4.0.1"
flake8-aaa = "0.12.2"
flake8-bandit = "3.0.0"
flake8-bugbear = "^22.4.25"
flake8-cognitive-complexity = "0.1.0"
flake8-comprehensions = "^3.10.0"
flake8-datetimez = "20.10.0"
flake8-functions = "0.0.7"
flake8-mutable = "^1.2.0"
flake8-pytest = "^1.3"
flake8-simplify = "0.19.2"
flake8-super = "0.1.3"
flake8-super-call = "^1.0.0"
flake8-tidy-imports = "^4.8.0"
flake8-tuple = "0.4.1"
bandit = "^1.7.4"
pep8-naming = "^0.13.0"
mastool = "^0.1.3"
# mccabe = "^0.7"

# Formatting
isort = "^5.10.1"
black = "^22.3.0"
docformatter = "^1.4"

# Typechecking
mypy = "^0.961"
types-toml = "^0.10.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
